Extracts more general settings into individual components
