+++ frontend/app/src/components/settings/general/BtcDerivationGapLimitSetting.vue
@@ -1,0 +1,39 @@
+<template>
+  <settings-option
+    #default="{ error, success, update }"
+    setting="btcDerivationGapLimit"
+    :error-message="$tc('general_settings.validation.btc_derivation_gap.error')"
+    :success-message="
+      limit =>
+        $tc('general_settings.validation.btc_derivation_gap.success', 0, {
+          limit
+        })
+    "
+  >
+    <v-text-field
+      v-model.number="btcDerivationGapLimit"
+      outlined
+      class="general-settings__fields__btc-derivation-gap"
+      :label="$t('general_settings.labels.btc_derivation_gap')"
+      type="number"
+      :success-messages="success"
+      :error-messages="error"
+      @change="update($event ? parseInt($event) : $event)"
+    />
+  </settings-option>
+</template>
+
+<script setup lang="ts">
+import { onMounted, ref } from '@vue/composition-api';
+import { get, set } from '@vueuse/core';
+import { useSettings } from '@/composables/settings';
+
+const btcDerivationGapLimit = ref<string>('20');
+
+const { generalSettings } = useSettings();
+
+onMounted(() => {
+  const settings = get(generalSettings);
+  set(btcDerivationGapLimit, settings.btcDerivationGapLimit.toString());
+});
+</script>
